Fix amount of network test loops, update comments.
I had a math error in the number of network test loops to perform. I'd originally said to test "5" at "10 sec", thinking that it would do its last test at 50 seconds on the clock. However the first test is at 0:00, so 5 tests would actually end at 0:40 instead of the desired 0:50. So increase this to 6, in order to get a full set of tests at every 10 seconds in each minute.

Also update the code comments to be clearer about the loop/retry behavior in the network tests.