Address issues #16 and #17.
Increase the length of the hysteresis to handle it better when the stream must be bounced due to a network problem. Add back in the pause after the bounce (previously commented out) to help ensure that it doesn't cry wolf immediately after the bounce. Change two quick-succession log messages into a single log message so that they don't cause the two log entries to appear in the wrong order in the Synology log.