Move SecondsToTime into helper functions, add WIP stream key code.
Move the SecondsToTime function into the Helper Functions file.

Update SecondsToTime so that it can display days as well as hours and minutes if needed.

Create a temporary WIP (work in progress) file that contains the example code for dealing with the YouTube stream key, which will eventually be used to fix issue #24 .