Work around issue #24 - Update the YouTube secret stream name/key if it changes.
Update the YouTube live stream name/key if they clobbered it.

Refactor code to allow for YouTube API authorization code to be in the Helper Functions shared file, so that it can be run from more than one script.

Update README with the new behavior.

Remove the temporary WIP code file.