Fix issues #27 and #28 - More robust stream key checking.
Fix for issue #27: Do not fixup the secret key if the stream's main title description does not match what we expect it to be.

Fix for issue #28: If there is a problem reading data from the YouTube API, don't always crash out of the program. In some cases it's important to continue running, for instance, if we still need to perform the network/uptime check.

Also: Updated some code comments.

Also: Made sure all script files were using unix LFs instead of Windows CRLFs.