Addressing issues #23 and #26 - Work in progress.
Enable the feature which uses the YouTube API to check the stream status.

Currently this runs in parallel with the mode which uses YouTube-DL to check the stream status. A failure in either the API method or the YouTube-DL method will consider the stream to be down.

If the API version works well over time, I will eventually remove the YouTube-DL method.