Correct case typo for videoIngestionStarved
Issue #23 - It's possible that the code that checks for "videoIngestionStarved" may have had a typo. I was going by something I found online which had a capital V, but I think the output from the actual API call might have the first character lowercase because the other similar variable, videoIngestionFasterThanRealtime, also starts with lower case, and I know that one works. The string check in bash is case-sensitive by default, I think so I have to fix this if I want accurate results.