Work-in-progress to fix stream privacyStatus
Begin work to address issue #37. Incomplete, the code does not work yet. Instead of fixing the settings of the current live stream, it creates a new upcoming event with those new settings.

Also: update the syntax of some IF statements which tested booleans, but which did not have brackets around them. They worked correctly before without the brackets, but the bad syntax set a dangerous precedent; the bracketless syntax only worked for "true" tests, not "false" tests. Update the syntax with brackets, so that all IF statements work both ways.