Fixed: Update the privacyStatus if YouTube clobbered it.
Fixes the worst (most frequent) offender of issue #37 - The privacyStatus of the YouTube stream. If it's not set to "public" (if YouTube clobbered and set it back to "unlisted"), then fix it back to "public".

There is potential for other values to get clobbered randomly by YouTube as well. If that issue recurs, I can expand this same code to perform a similar fix on the other clobbered data, at some point in the future.