Make sure stream splits are recorded only when long bounces are used.
I realized that my last checkin might regress in certain situations. If I did a short bounce, which didn't split the stream, but wrote out the stream start time anyway, then on an otherwise "good" summer day with a pre-midday stream blip, it might fail to split the stream at the split point because it thought the "blip" would have split it already.

This checkin should theoretically fix that issue and only record a StreamStartTime at the appropriate moments:
- When stream is started in the morning.
- When stream is split with a long bounce only (not a short bounce).