Update cleanup behavior and frequency.
The CrowCamCleanup is now being used for saving off the video data file, and I want that done more frequently. There isn't an "hourly" option in Synology Tash Scheduler, so I'm setting it to every 30 minutes.

Change the way the video data file gets written, so that it only gets written if the program passes all its validations.