Work in progress - Add custom playlist handling to Cleanup.
Work in progress to add custom playlist handling to CrowCamCleanup.sh.

Instead of scanning and cleaning only the "uploads" playlist, it will scan and clean either "uploads" or a specified playlist.

This has a bug still, that I'm working on. When cleaning a specified playlist, for unlisted videos, it retrieves "publishedAt" variables for the date the item was added to the playlist, not the date the video was uploaded. This messes up the crowcam-videodata file that I'm using somewhere else. Still not sure how to fix that.