Retrieve actual start and end time at the cost of 2 quota per video.
Because the "publishedAt" variable is in bad shape, instead, burn some quota to retrieve the actual start and end time of every video. Reduce the number of times per day that this gets queried by giving instructions to set it to every 6 hours instead of every hour.

To do: When this is stable, start using this variable for the actual deletion, and skip deletion of any one which doesn't have an actual start or end time.