Invalidate cache for young 0Zulu videos too.
This should theoretically address the last of the issue#58 stuff. If a video LOOKS like an uploaded video (no end time, timestamp of 0Zulu), then leave it in the cache UNLESS it's young. If it's super young, then keep invalidating the cache until it ether gets old or it gets an end time.