Improve network test
I ran into a situation where the network test failed. The cable modem was down, so the DHCP/DNS router took over and hijacked requests to "google.com" and put its own web page up in its place.

Switch to pinging the 8.8.8.8 google DNS server as the test, instead of "google.com" to ensure that it's testing something valid.