Fix error logging to midday instead of midnight
One of the earlier logging improvements was to only Synology-log an error about YouTube-DL during the midnight hour, so that problems with their website didn't spam the log. However, my choice of doing it only at midnight was foolish: since the stream is down at midnight, the keep-alive doesn't run at midnight. Fix it by changing the logging to midday so that it works as intended. This is continued secondary effects of issue #66.