Experimentally disable the hourly quick bounce.
Working with issue #14 - Synology is working with me on an experimental patch to fix the audio desynchronization. I am temporarily disabling the hourly bounce, so that if the issue recurs with the patch in place, I will more readily be able to witness the issue. If the issue never recurs, this setting can simply be left turned off.