Fix livestream issues on the YouTube side.
On 2020-09-10 there seemed to be some changes to the live stream features of YouTube, possibly part of their shift from the "old" version of Studio to the "new" version of Studio. Some changes to the code were needed to cover this, in particular, some of the variables are no longer retrieved, but luckily those didn't seem to be needed.