Change to broadcastType=all
I'm not sure if this is caused by a change in YouTube, or if I can't set up my account right, or what. But I can't seem to get the script to work with broadcastType=persistent any more and the only way these three code sections now work is broadcastType=all. Experimenting with this.