Fixed OAuth code retrieval
GitHub issue #69 - Partially fix the problem with the retrieval of the OAuth code due to deprecation of the OOB flow. This does not completely fix GitHub issue #69 but it fixes part of it. The problem is that the end user must now create their own localhost web page to display the code. I still need to write up some instructions on how to do this part.