Transfer old playlist to new, speed up Cleanup.
Github issue #69 - Some fixes to ensure that the CrowCamCleanup was working properly and also correctly delivering data to my personal side project that updates my web site. Note that this does not fix GitHub issue #69 yet, it's just part of it.

GitHub Issue #72 - Improve the speed of CrowCamCleanup.sh by changing the way one of the big JSON parsing loops works. This is a significant increase. Leaving #72 open because I think it can be made even better in the future.