Bugfix GetSunriseSunsetTimeFromGoogle
When Google changed the way they display sunrise/sunset times, I had to change the parsing code. The new parsing code had some bugs in it which meant it only worked in the perfect golden path. There were two possibilities that weren't taken into account: What happens when more than one time is found in the google results (answer: it screws it up, so you have to add -m 1 to the grep), and what happens if there are no results found (answer: it returns a space instead of a null string, screwing it up, so you have to return two different things depending on what it found).