Final fix for issue #69 ?
Betteridge's law of headlines is an adage that states: "Any headline that ends in a question mark can be answered by the word no." It is named after Ian Betteridge, a British technology journalist who wrote about it in 2009, although the principle is much older. I hope that's not the case here. I believe that this checkin cleans up the last major problem in the fixes for issue #69.