Fix sunrise/sunset retrieval bug.
Fix issue #81 - To fix this, I had to turn on extended regex (-E), and instead of searching for a single weird unicode character ("."), I had to search for 1-3 characters (".{1,3}"). This was only necessary when running under the Synology Task Scheduler; the problem didn't manifest itself when running on the Synology at the SSH shell prompt.