Don't check video title when checking to fix stream key
Fix issue #82 which essentially reverts the fix to issue #27. Don't check the stream title when we are checking if the stream key needs to be fixed. This allows me to rename a video while it's still livestreaming, and will stop the errors appearing in the log when I do that.