Fix Sunrise/Sunset time retrieval
This should be the final fix for issues #81 and #84 hopefully. Set the "LC_ALL" variable if it's not already set, which allows for parsing of single unicode characters. Then update the time parsing code to be more specific about the things it's retrieving, and allow it to get upper/lowercase AM/PM with and without periods.