Check and fix category, playlist, and thumbnail
Fix for issue #85 - Fix the categoryId, the thumbnail, and the playlist of the currently live-streaming video if those things happen to be bad. The thumbnail is not actively checked, but if either the category or the playlist are bad, then re-upload the thumbnail too, because it's likely to also be bad too.