Create a new stream if the local key doesn't match the YouTube key
Work around GitHub issue #91. Synology API has a bug where, if you have a stream key that starts with four decimal digits then the API fails to set the key in the Live Broadcast widget. Work around this by rechecking the key after setting it. If the key still fails, then simply create a new stream from scratch and hope that it will have a key that starts with a combination of letters and numbers.

This has a side effect of leaving an orphaned "upcoming live broadcast" sitting in the YouTube channel, but that's better than no stream at all.