Use doublequotes around the stream key
Additional workaround for GitHub issue #91 - Add doublequotes around the stream key when setting the value via the Synology API. This was suggested by Synology support. I'm using the URL-escaped version %22 instead of an actual doublequote, since I'm already using doublequotes as the main string delimiter for the URL, but the %22 also works equally well to fix the problem.