Remove synoservicectl checks
GitHub Issue #77 - According to contributor @quog, the command "synoservicectl" has been removed from Synology DSM version 7.x.

I only used synoservicectl to check the running status of the Surveillance Station service, and I only checked it for making sure that the user met the requirements of running Surveillance Station. So it is safe to completely remove the code which references synoservicectl. Now the code will not check to see if the user is running Surveillance Station.

This change removes all code that uses the following commands and variables:

serviceName
synoservicectl
IsServiceUp
currentServiceState