Authenticate with DSM API differently
GitHub issue #77: Fix the DSM API authentication login so that it works on both DSM6 and DSM7. Detect if the NAS is running DSM v6/earlier or v7/later by trying to do the api queries indicated by their authentication docs. Doesn't quite work the way they said it would work because they moved the commands between different CGI files from DSM6 to DSM7. Use this to our advantage to detect which side of that line we're on, and change the authentication API string depending on what we found out.