Implement SunriseSunset.io
Fix for GitHub issue #96 - Add SunriseSunset.io as an additional way to retrieve the sunrise and sunset times. I had initially rejected using SunriseSunset.io about a year ago, and still rejected it a few weeks ago when fixing issue #95. But in the end, I agree with Damian that, philosophically, an API is always better than a screen scrape. I'm using it only as a secondary method at the moment, only if the screen scrape function fails, and my reasons for doing it that way are given in the function description.