Fix bug #92 and hopefully #93
This should definitively fix bug #92 by ensuring that my sed/grep/cut trick for extracting JSON variables always works.

The problem was that I was correctly "sed"ing for the variable name with the quotes, but I was incorrectly "grep"ing without the quotes. This meant that sometimes if there was an earlier string that happened to include the variable name (such as a key or guid string randomly containing the letters "id") then it would get the first variable in the file rather than the variable I was looking for.

This should fix general type of issue globally for all of the code in CrowCam. There is another bug that I encountered which might also be caused by this same issue, so we should see if it get fixed by this too: bug #93