Fix sunrise time on DST change day, bug #97
Address bug #97 where, on the Sunday morning after the DST change, the camera turned on at the wrong hour. Fix it by changing the timing windows where it retrieves the sunrise/sunset times from a web site. This is to ensure that it retrieves the time of the UPCOMING sunrise rather than the one that happened yesterday.